repo cleanup part1
