+++ repo/repo.go
@@ -3,23 +3,9 @@
 package repo
 
 import (
-	"log"
-	"os"
 	"path/filepath"
-	"regexp"
 
-	"github.com/dgraph-io/badger"
 	"github.com/pkg/errors"
-	"go.cryptoscope.co/librarian"
-	libmkv "go.cryptoscope.co/librarian/mkv"
-	"go.cryptoscope.co/margaret"
-	"go.cryptoscope.co/margaret/codec/msgpack"
-	"go.cryptoscope.co/margaret/multilog"
-	multibadger "go.cryptoscope.co/margaret/multilog/badger"
-	"go.cryptoscope.co/margaret/multilog/roaring"
-	multifs "go.cryptoscope.co/margaret/multilog/roaring/fs"
-	multimkv "go.cryptoscope.co/margaret/multilog/roaring/mkv"
-	"modernc.org/kv"
 
 	"go.cryptoscope.co/ssb"
 	"go.cryptoscope.co/ssb/blobstore"
@@ -40,225 +26,7 @@ func (r repo) GetPath(rel ...string) string {
 	return filepath.Join(append([]string{r.basePath}, rel...)...)
 }
 
-const PrefixMultiLog = "sublogs"
-
-// OpenBadgerMultiLog uses the repo to determine the paths where to finds the multilog with given name and opens it.
-//
-// Exposes the badger db for 100% hackability. This will go away in future versions!
-// badger + librarian as index
-func OpenBadgerMultiLog(r Interface, name string, f multilog.Func) (multilog.MultiLog, librarian.SinkIndex, error) {
-
-	dbPath := r.GetPath(PrefixMultiLog, name, "db")
-	err := os.MkdirAll(dbPath, 0700)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "mkdir error for %q", dbPath)
-	}
-
-	db, err := badger.Open(badgerOpts(dbPath))
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "db/idx: badger failed to open")
-	}
-
-	mlog := multibadger.New(db, msgpack.New(margaret.BaseSeq(0)))
-
-	statePath := r.GetPath(PrefixMultiLog, name, "state.json")
-	mode := os.O_RDWR | os.O_EXCL
-	if _, err := os.Stat(statePath); os.IsNotExist(err) {
-		mode |= os.O_CREATE
-	}
-	idxStateFile, err := os.OpenFile(statePath, mode, 0700)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "error opening state file")
-	}
-
-	mlogSink := multilog.NewSink(idxStateFile, mlog, f)
-
-	return mlog, mlogSink, nil
-}
-
-func OpenFileSystemMultiLog(r Interface, name string, f multilog.Func) (*roaring.MultiLog, librarian.SinkIndex, error) {
-	dbPath := r.GetPath(PrefixMultiLog, name, "fs-bitmaps")
-	err := os.MkdirAll(dbPath, 0700)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "mkdir error for %q", dbPath)
-	}
-	mlog, err := multifs.NewMultiLog(dbPath)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "open error for %q", dbPath)
-	}
-
-	if err := mlog.CompressAll(); err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to compress db")
-	}
-
-	// todo: save the current state in the multilog
-	statePath := r.GetPath(PrefixMultiLog, name, "state_mkv.json")
-	mode := os.O_RDWR | os.O_EXCL
-	if _, err := os.Stat(statePath); os.IsNotExist(err) {
-		mode |= os.O_CREATE
-	}
-	idxStateFile, err := os.OpenFile(statePath, mode, 0700)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "error opening state file")
-	}
-
-	mlogSink := multilog.NewSink(idxStateFile, mlog, f)
-
-	return mlog, mlogSink, nil
-
-}
-
-func OpenMultiLog(r Interface, name string, f multilog.Func) (multilog.MultiLog, librarian.SinkIndex, error) {
-
-	dbPath := r.GetPath(PrefixMultiLog, name, "roaring")
-	err := os.MkdirAll(dbPath, 0700)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "mkdir error for %q", dbPath)
-	}
-
-	mkvPath := filepath.Join(dbPath, "mkv")
-	mlog, err := multimkv.NewMultiLog(mkvPath)
-	if err != nil {
-		// yuk..
-		if !isLockFileExistsErr(err) {
-			// delete it if we cant recover it
-			os.RemoveAll(dbPath)
-			return nil, nil, errors.Wrapf(err, "not a lockfile problem - deleting index")
-		}
-		if err := cleanupLockFiles(dbPath); err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to recover lockfiles")
-
-		}
-		mlog, err = multimkv.NewMultiLog(mkvPath)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to open roaring db")
-		}
-	}
-
-	if err := mlog.CompressAll(); err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to compress db")
-	}
-
-	// todo: save the current state in the multilog
-	statePath := r.GetPath(PrefixMultiLog, name, "state_mkv.json")
-	mode := os.O_RDWR | os.O_EXCL
-	if _, err := os.Stat(statePath); os.IsNotExist(err) {
-		mode |= os.O_CREATE
-	}
-	idxStateFile, err := os.OpenFile(statePath, mode, 0700)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "error opening state file")
-	}
-
-	mlogSink := multilog.NewSink(idxStateFile, mlog, f)
-
-	return mlog, mlogSink, nil
-}
-
-func cleanupLockFiles(root string) error {
-	return filepath.Walk(root, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-		name := filepath.Base(path)
-		if info.Size() == 0 && len(name) == 41 && name[0] == '.' {
-			log.Println("dropping empty lockflile", path)
-			if err := os.Remove(path); err != nil {
-				return errors.Wrapf(err, "failed to remove %s", name)
-			}
-		}
-		return nil
-	})
-}
-
-const PrefixIndex = "indexes"
-
-func OpenIndex(r Interface, name string, f func(librarian.SeqSetterIndex) librarian.SinkIndex) (librarian.Index, librarian.SinkIndex, error) {
-	pth := r.GetPath(PrefixIndex, name, "mkv")
-	err := os.MkdirAll(pth, 0700)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "openIndex: error making index directory")
-	}
-
-	db, err := OpenMKV(pth)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "openIndex: failed to open MKV database")
-	}
-
-	idx := libmkv.NewIndex(db, margaret.BaseSeq(0))
-	return idx, f(idx), nil
-}
-
-func OpenMKV(pth string) (*kv.DB, error) {
-	opts := &kv.Options{}
-	os.MkdirAll(pth, 0700)
-	dbname := filepath.Join(pth, "idx.mkv")
-	var db *kv.DB
-	_, err := os.Stat(dbname)
-	if os.IsNotExist(err) {
-		db, err = kv.Create(dbname, opts)
-		if err != nil {
-			return nil, errors.Wrap(err, "failed to create mkv")
-		}
-	} else if err != nil {
-		return nil, errors.Wrap(err, "failed to stat path location")
-	} else {
-		db, err = kv.Open(dbname, opts)
-		if err != nil {
-
-			if !isLockFileExistsErr(err) {
-				return nil, err
-			}
-			if err := cleanupLockFiles(pth); err != nil {
-				return nil, errors.Wrapf(err, "failed to recover lockfiles")
-			}
-			db, err = kv.Open(dbname, opts)
-			if err != nil {
-				return nil, errors.Wrap(err, "failed to open mkv")
-			}
-		}
-	}
-	return db, nil
-}
-
-type LibrarianIndexCreater func(*badger.DB) (librarian.SeqSetterIndex, librarian.SinkIndex)
-
-func OpenBadgerIndex(r Interface, name string, f LibrarianIndexCreater) (*badger.DB, librarian.SeqSetterIndex, librarian.SinkIndex, error) {
-	pth := r.GetPath(PrefixIndex, name, "db")
-	err := os.MkdirAll(pth, 0700)
-	if err != nil {
-		return nil, nil, nil, errors.Wrap(err, "error making index directory")
-	}
-
-	db, err := badger.Open(badgerOpts(pth))
-	if err != nil {
-		return nil, nil, nil, errors.Wrap(err, "db/idx: badger failed to open")
-	}
-
-	idx, sinkidx := f(db)
-
-	return db, idx, sinkidx, nil
-}
-
 func OpenBlobStore(r Interface) (ssb.BlobStore, error) {
 	bs, err := blobstore.New(r.GetPath("blobs"))
 	return bs, errors.Wrap(err, "error opening blob store")
-}
-
-var lockFileExistsRe = regexp.MustCompile(`cannot access DB \"(.*)\": lock file \"(.*)\" exists`)
-
-func isLockFileExistsErr(err error) bool {
-	log.Println("TODO: check process isn't running")
-	if err == nil {
-		return false
-	}
-	errStr := errors.Cause(err).Error()
-	if !lockFileExistsRe.MatchString(errStr) {
-		return false
-	}
-	matches := lockFileExistsRe.FindStringSubmatch(errStr)
-	if len(matches) == 3 {
-		return true
-	}
-	return false
 }
